util: move deadlock into mocktikv (#24103)
